Do not unregister IRCBot from message router
It seems IRCBot did not receive the 'you are now identified for' notice
message after reconnecting. Perhaps this is why the reconnecting itself
seemed to work properly but no channels were joined. There should be no
significant overhead to keep IRCBot as a message listener, so do not
unregister it.