Add Menu plugin, for handling per-user entry browsing
This is generalization of what is used in EDICT/Daijirin.
Menu is considered to be a tree of MenuNode-s, each of which has
"description" (to be displayed alongside other choices
in an enumerated list), and "enter" procedure, which performs whatever
action necessary, and also returns sub-menu array, if any.

Menu-s are kept on per-user/channel basis, and are subject to expiration.