Add a script for ruby to see all bundler gems.
The reasons why this is needed to run the bot are not known yet:
rubyeb19 requires this to be seen by ruby, but not all git gems do.
However, there is no issue loading rubyeb19 from `bundle exec ruby`,
or when using `require 'bundler/setup'`, which is an equivalent,
hence further investigation is required.

Typical use case: ./bundlexec.sh ./bot.rb